+++ Cargo.toml
@@ -19,23 +19,23 @@ version = "0.1.0"
 edition = "2018"
 
 [dependencies]
-byteorder = "^0.5.3"
-cgmath = "0.16.0"
-collision = "0.18.0"
+byteorder = "1.3.2"
+cgmath = "0.17.0"
+collision = "0.20.1"
 crossbeam = "0.7.1"
-error-chain = "0.11.0"
+error-chain = "0.12.1"
 fnv = "1.0.6"
-libc = "0.2.43"
-nav-types = "0.4.0"
-num = "0.1.36"
-num-traits = "0.1.36"
+libc = "0.2.58"
+nav-types = "0.4.1"
+num = "0.2.0"
+num-traits = "0.2.8"
 objekt = "0.1.2"
-pbr = "1.0.0-alpha.1"
-protobuf = "2.0.0"
+pbr = "1.0.1"
+protobuf = "2.7.0"
 s2 = "0.0.8"
-scoped-pool = "^0.1"
-structopt = "0.2"
-walkdir = "^0.1.5"
+scoped-pool = "1.0.0"
+structopt = "0.2.18"
+walkdir = "2.2.8"
 
 [dependencies.point_viewer_proto_rust]
 path = "point_viewer_proto_rust"
