+++ octree_web_viewer/Cargo.toml
@@ -21,8 +21,8 @@ edition = "2018"
 [dependencies]
 actix = "0.7.9"
 actix-web = "0.7.18"
-byteorder = "^0.5.3"
-cgmath = "0.16.0"
+byteorder = "1.3.2"
+cgmath = "0.17.0"
 clap = "2.32.0"
 crossbeam-utils = "0.6.5"
 failure = "0.1.5"
@@ -30,7 +30,7 @@ futures = "0.1.25"
 json = "0.11.3"
 serde = "1.0.89"
 serde_derive = "1.0.89"
-structopt = "0.2.15"
+structopt = "0.2.18"
 time = "0.1.42"
 
 [dependencies.point_viewer]
