+++ point_viewer_grpc/Cargo.toml
@@ -19,15 +19,15 @@ version = "0.1.0"
 edition = "2018"
 
 [dependencies]
-collision = "0.18.0"
-cgmath = "0.16.0"
+collision = "0.20.1"
+cgmath = "0.17.0"
 clap = "2.32.0"
 crossbeam-channel = "0.3.8"
 ctrlc = "3.1.2"
 futures = "0.1.17"
 grpcio = "0.4"
-protobuf = "2.0.0"
-scoped-pool = "^0.1"
+protobuf = "2.7.0"
+scoped-pool = "1.0.0"
 
 [dependencies.point_viewer]
 path = ".."
