+++ point_viewer_grpc/src/lib.rs
@@ -19,9 +19,9 @@ use futures::{Future, Stream};
 use grpcio::{ChannelBuilder, EnvBuilder};
 use point_viewer::color::Color;
 use point_viewer::errors::*;
-use point_viewer::octree::{NodeId, NodeLayer, Octree, OctreeDataProvider};
+use point_viewer::octree::{NodeId, Octree, OctreeDataProvider};
 use point_viewer::proto::Meta;
-use point_viewer::Point;
+use point_viewer::{NodeLayer, Point};
 pub use point_viewer_grpc_proto_rust::proto;
 pub use point_viewer_grpc_proto_rust::proto_grpc;
 use std::collections::HashMap;
