+++ point_viewer_grpc/src/service.rs
@@ -63,7 +63,7 @@ fn send_fail<T>(ctx: &RpcContext, sink: UnarySink<T>, err_str: String) {
 
 impl proto_grpc::Octree for OctreeService {
     fn get_meta(
-        &self,
+        &mut self,
         ctx: RpcContext,
         req: proto::GetMetaRequest,
         sink: UnarySink<proto::GetMetaReply>,
@@ -81,7 +81,7 @@ impl proto_grpc::Octree for OctreeService {
     }
 
     fn get_node_data(
-        &self,
+        &mut self,
         ctx: RpcContext,
         req: proto::GetNodeDataRequest,
         sink: UnarySink<proto::GetNodeDataReply>,
@@ -111,7 +111,7 @@ impl proto_grpc::Octree for OctreeService {
     }
 
     fn get_points_in_frustum(
-        &self,
+        &mut self,
         ctx: RpcContext,
         req: proto::GetPointsInFrustumRequest,
         resp: ServerStreamingSink<proto::PointsReply>,
@@ -146,7 +146,7 @@ impl proto_grpc::Octree for OctreeService {
     }
 
     fn get_points_in_box(
-        &self,
+        &mut self,
         ctx: RpcContext,
         req: proto::GetPointsInBoxRequest,
         resp: ServerStreamingSink<proto::PointsReply>,
@@ -168,7 +168,7 @@ impl proto_grpc::Octree for OctreeService {
     }
 
     fn get_all_points(
-        &self,
+        &mut self,
         ctx: RpcContext,
         req: proto::GetAllPointsRequest,
         resp: ServerStreamingSink<proto::PointsReply>,
@@ -181,7 +181,7 @@ impl proto_grpc::Octree for OctreeService {
     }
 
     fn get_points_in_oriented_beam(
-        &self,
+        &mut self,
         ctx: RpcContext,
         req: proto::GetPointsInOrientedBeamRequest,
         resp: ServerStreamingSink<proto::PointsReply>,
