+++ point_viewer_grpc_proto_rust/Cargo.toml
@@ -19,11 +19,11 @@ authors = ["Holger Rapp <hrapp@lyft.com>"]
 edition = "2018"
 
 [build-dependencies]
-protoc-rust = "2.0.0"
-protoc-grpcio = "0.2.1"
+protoc-rust = "2.7.0"
+protoc-grpcio = "1.0.2"
 
 [dependencies]
-protobuf = "2.0.0"
+protobuf = "2.7.0"
 grpcio = "0.4"
 futures = "0.1.17"
