+++ xray/Cargo.toml
@@ -5,20 +5,20 @@ authors = ["Holger Rapp <hrapp@lyft.com>"]
 edition = "2018"
 
 [dependencies]
-fnv = "1.0.6"
-cgmath = "0.16.0"
+cgmath = "0.17.0"
 clap = "2.32.0"
-collision = "0.18.0"
+collision = "0.20.1"
+fnv = "1.0.6"
 image = "0.20.0"
 iron = "^0.3.0"
-num = "0.1.36"
-protobuf = "2.0.0"
+num = "0.2.0"
+protobuf = "2.7.0"
 router = "^0.1.1"
-scoped-pool = "^0.1"
+scoped-pool = "1.0.0"
 serde = "1.0.9"
 serde_json = "1.0.9"
 serde_derive = "1.0.9"
-structopt = "0.2"
+structopt = "0.2.18"
 urlencoded = "^0.3.0"
 streaming-stats = "0.2.2"
