+++ xray_proto_rust/Cargo.toml
@@ -19,7 +19,7 @@ authors = ["Holger Rapp <hrapp@lyft.com>"]
 edition = "2018"
 
 [build-dependencies]
-protoc-rust = "2.0.0"
+protoc-rust = "2.7.0"
 
 [dependencies]
-protobuf = "2.0.0"
+protobuf = "2.7.0"
