+++ xray_proto_rust/build.rs
@@ -3,11 +3,6 @@ use std::fs::File;
 use std::io::{Read, Write};
 use std::path::Path;
 
-// This is a workaround for https://github.com/stepancheg/rust-protobuf/issues/331
-fn replace_clippy(contents: String) -> String {
-    contents.replace("#![allow(clippy)]", "#![allow(clippy::all)]")
-}
-
 fn main() {
     println!("cargo:rerun-if-changed=src/proto.proto");
 
@@ -34,7 +29,7 @@ fn main() {
         .unwrap()
         .read_to_string(&mut contents)
         .unwrap();
-    let new_contents = format!("pub mod proto {{\n{}\n}}", replace_clippy(contents));
+    let new_contents = format!("pub mod proto {{\n{}\n}}", contents);
 
     File::create(&proto_path)
         .unwrap()
